refactor(reporter): isolate provider telemetry
